requirements.txt
google-generativeai
fastapi
uvicorn[standard]
httpx
pydantic
python-dotenv
streamlit